[structure] Major restructure and code organization
