+++ src/actions/api.js
@@ -1,2 +1,3 @@
+export const API_URL = 'https://guts-dev.herokuapp.com'
 export const API_FETCH = 'API_FETCH'
 export const API_RECIEVE = 'API_RECIEVE'
