+++ src/index.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { Provider } from 'react-redux'
 import { render } from 'react-dom'
-import configureStore from './store/configureStore'
-import App from './App'
+import configureStore from './store/configure'
+import App from 'modules/app'
 
 const store = configureStore()
 
 render(
   <Provider store={store}>
-      <App />
+    <App />
   </Provider>,
   document.getElementById('root')
 )
