+++ src/modules/tournaments/index.js
@@ -1,0 +1,3 @@
+import TournamentList from './components/tournament_list'
+
+export default TournamentList
