+++ src/store/configure.js
@@ -1,5 +1,5 @@
 import { createStore, applyMiddleware } from 'redux'
-import rootReducer from '../reducers/rootReducer'
+import rootReducer from './rootReducer'
 import thunk from 'redux-thunk'
 
 export default function configureStore() {
